Add files via upload
Our poject consists in smart traffic lights. We propous a system which takes traffic information and sends it to a microcontroller that acording with the traffic state decides how many time traffic light must be letting the cars flow, or stopping it. This system does not use cameras, which makes its implementation less expensive.